refactor: Move TimePoint into core package
