+++ src/main/java/seedu/address/commons/core/timepoint/TimePointComparator.java
@@ -1,4 +1,4 @@
-package seedu.address.model.timepoint;
+package seedu.address.commons.core.timepoint;
 
 import java.util.Comparator;
