+++ src/test/java/seedu/address/commons/core/timepoint/TimeComparatorTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.timepoint;
+package seedu.address.commons.core.timepoint;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
